requirements.txt
click==8.1.7
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
xgboost==2.0.3
tensorflow==2.13.1
keras==2.13.1